2010-02-08 23:18 UTC+0100 Viktor Szakats (harbour.01 syenar.hu)   * src/rtl/gtwin/gtwin.c   * include/hbrddcdx.h   * contrib/hbct/screen2.c   * contrib/hbct/ct.h   * contrib/xhb/freadlin.c   * contrib/gtwvg/gtwvg.h     ! Cleaned remaining legacy types after verification with grep script.

+++ harbour/contrib/xhb/freadlin.c
@@ -129,7 +129,7 @@ char * hb_fsReadLine( HB_FHANDLE hFileHandle, HB_ISIZ * plBuffLen, const char **
             pBuff[ *plBuffLen ] = '\0';
 
             /* Set handle pointer in the end of the line */
-            hb_fsSeek( hFileHandle, (((lRead-(( LONG ) iPos )))*-1)+1, FS_RELATIVE );
+            hb_fsSeek( hFileHandle, ((lRead-iPos)*-1)+1, FS_RELATIVE );
 
             return pBuff;
          }
